+++ spec/terraform_plan_spec.rb
@@ -41,6 +41,18 @@ describe TerraformLandscape::TerraformPlan do
       OUT
     end
 
+    context 'when output contains a single resource with no attributes' do
+      let(:terraform_output) { normalize_indent(<<-TXT) }
+        - some_resource_type.some_resource_name
+
+      TXT
+
+      it { should == normalize_indent(<<-OUT) }
+        - some_resource_type.some_resource_name
+
+      OUT
+    end
+
     context 'when output contains a single resource with multiple attributes' do
       let(:terraform_output) { normalize_indent(<<-TXT) }
         ~ some_resource_type.some_resource_name